Refactor feature configuration and extraction
